+++ src/methods/entries/__tests__/entries.test.js
@@ -1,20 +1,17 @@
 import { entries } from '../../..';
 
 describe('entries', () => {
-  it('works with Map', () => {
-    const mapEntries: Array<[string, string]> = [['foo', 'foo'], ['bar', 'bar']];
-    const map = new Map(mapEntries);
-    expect(Array.from(entries(map))).toEqual(mapEntries);
-  });
   it('works with Objects', () => {
     const i = entries({ '1': 1, '2': 2 });
     expect(Array.from(i)).toEqual([['1', 1], ['2', 2]]);
   });
+
   it('can be reused', () => {
     const i = entries({ '1': 1, '2': 2 });
     expect(Array.from(i)).toEqual([['1', 1], ['2', 2]]);
     expect(Array.from(i)).toEqual([['1', 1], ['2', 2]]);
   });
+
   it('works with null', () => {
     const i = entries(null);
     expect(Array.from(i)).toEqual([]);