+++ src/methods/entries/entries.mjs
@@ -2,18 +2,17 @@ import { empty, wrapWithMethodIterable } from '../../internal/iterable';
 
 const { hasOwnProperty } = Object.prototype;
 
-export function* entries(entriesable) {
-  if (entriesable == null) {
+export function* entries(obj) {
+  if (obj == null) {
     return empty();
-  } else if (typeof entriesable.entries === 'function') {
-    yield* entriesable.entries();
-  } else if (typeof entriesable === 'object') {
-    // pojo
-    for (const key in entriesable) {
-      if (hasOwnProperty.call(entriesable, key)) {
-        yield [key, entriesable[key]];
+  } else if (typeof obj === 'object') {
+    for (const key in obj) {
+      if (hasOwnProperty.call(obj, key)) {
+        yield [key, obj[key]];
       }
     }
+  } else {
+    throw new Error('the argument to entries() was not an object, null, or undefined');
   }
 }