+++ src/methods/keys/__tests__/keys.auto.spec.ts
@@ -8,10 +8,6 @@
 
 import { keys } from '../../..';
 describe('keys', () => {
-  it('works with Map', () => {
-    const map = new Map([['foo', 'bar'], ['fox', 'far']]);
-    expect(Array.from(keys(map))).toEqual(Array.from(map.keys()));
-  });
   it('works with Objects', () => {
     const i = keys({
       '1': 1,