+++ src/methods/values/values.d.ts
@@ -1,7 +1,5 @@
 import { Iterable, ResultIterable } from '../../types/iterable';
 
-declare function values<V>(valuesable: { [id: string]: V } | null | undefined): ResultIterable<V>;
-
-declare function values<V>(valuesable: { values(): Iterable<V> }): ResultIterable<V>;
+declare function values<V>(obj: { [id: string]: V } | null | undefined): ResultIterable<V>;
 
 export default values;