+++ src/methods/wrap-keys/wrap-keys.d.ts
@@ -1,0 +1,7 @@
+import { Iterable, ResultIterable } from '../../types/iterable';
+
+declare function wrapKeys<K>(
+  keysable: { keys(): Iterable<K> } | null | undefined,
+): ResultIterable<K>;
+
+export default wrapKeys;